ByteStringUtil class 	- I put a  "public static final Charset STRING_ENCODING = Charset.forName("UTF-8");" in it for use in the entire project. This should be whenever a String object deals with byte[] in anyway... like "String.getBytes(enc)" or "new String(byte[], enc)"
	- method "strcmp(CharSeq, byte[], int)" and "strcmp(CharSeq, ByteBuffer)"... used the const Charset defined above when turning String to byte[]
	- method "append(CharSeq, int, byte[], int)" and "append(DataOutput, CharSeq, int)"... same as above
	
CompactString class
	- removed local definition of ENCODING and used the one now defined in ByteStringUtil class (+ removed UnsupportedCharsetException handling since now unecessary)
	
ReplazableString class
	- same as with CompactString
	
IOUtil class
	- method writeLine... used the STRING_ENCODING Charset from ByteStringUtil class when turning String to byte[]

DictionarySectionDeflate class
	- load method... used the STRING_ENCODING Charset from ByteStringUtil class when turning String to byte[]
	
DictionarySectionJDBM class
	- all the DEBUG write-outs removed, code now as should be and works perfectly
	
---------------------------------------------------------------

I found the bug in the JDBM library and fixed it (had to do with String encoding). I build a new JAR from the fixed code and removed the old one. The new one is called: "JDBM-3.0+_myBuild+src", the "+" indicating that I changed the original code and "myBuild+src" indicating that I made the jar myself from the source and attached the source in it (so it can be accessed through Eclipse by ctrl+left_mouse_click on class/interface/method definition).

The JDBM version should now work perfectly, and the whole project as well, inependent of the platform it is run on (previously was dependent because of platform default charset use).

git-svn-id: https://hdt-java.googlecode.com/svn@48 79d02f10-9824-5a6d-8155-d2f4b3853b0f