TriplesSet set to deprecated! (and CacheCalculator writes out a warning message if TriplesSet used with on-disc dictionary)
This is because TriplesSet takes a lot of memory I forgot to count in, and I'm not sure how to calculate it depending on the amount of elements. The extra memory is of course the tree structure holding the TripleID's and I'm almost sure that each TreeNode is bigger than a TripleID. So anyway... my not being able to calculate the memory imprint is not that important because... TriplesList is better anyway (in almost all aspectes) so it should be used.
I left the TriplesSet there because no point in deleting it, it can still be useful, but I deprecated it's constructor so everybody that uses it knows :)

git-svn-id: https://hdt-java.googlecode.com/svn@80 79d02f10-9824-5a6d-8155-d2f4b3853b0f